corelens: split out v2 modules
